# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Library Management API",
      "description": "API for managing a library of books",
      "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
      "/books": {
        "get": {
          "summary": "Get all books",
          "responses": {
            "200": {
              "description": "A list of books",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Book"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new book",
          "parameters": [
            {
              "name": "book",
              "in": "body",
              "description": "Book to add",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Book created"
            }
          }
        }
      },
      "/books/{isbn}": {
        "get": {
          "summary": "Get a book by ISBN",
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Book details",
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        },
        "put": {
          "summary": "Update a book",
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "book",
              "in": "body",
              "description": "Book details to update",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Updated book"
            }
          }
        },
        "delete": {
          "summary": "Delete a book",
          "parameters": [
            {
              "name": "isbn",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "204": {
              "description": "Book deleted"
            }
          }
        }
      }
    },
    "definitions": {
      "Book": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "genre": {
            "type": "string"
          }
        }
      }
    }
  }
  